# Improve performance of ChatChannel and ChatMessage equality