### 🔗 Issue Links

Related to [PBE-5004](https://stream-io.atlassian.net/browse/PBE-5004)

### 🎯 Goal

* Reduce message and channel copies when calling `==`

### 📝 Summary

Time profiler shows in the SwiftUI SDK that when doing `==` checks a lot of time is spent in making copies of messages and channels. Changing the `==` to use guard let takes away a lot of this overhead.

### 🎨 Showcase

LazyCachedMapCollection's `==` is reduced from 118 ms > 11 ms  **>10** times improvement 

| Before | After |
| ------ | ----- |
|  <img width="951" alt="Before" src="https://github.com/user-attachments/assets/68f65b42-04e3-48c8-a3de-c5e9fb48cdcd">   |  <img width="770" alt="After" src="https://github.com/user-attachments/assets/ad687756-a11a-46a0-a57d-bcf1c5a634ee">  |

### 🧪 Manual Testing Notes

None

### ☑️ Contributor Checklist

- [x] I have signed the [Stream CLA](https://docs.google.com/forms/d/e/1FAIpQLScFKsKkAJI7mhCr7K9rEIOpqIDThrWxuvxnwUq2XkHyG154vQ/viewform) (required)
- [x] This change should be manually QAed
- [ ] Changelog is updated with client-facing changes
- [ ] Changelog is updated with new localization keys
- [ ] New code is covered by unit tests
- [ ] Comparison screenshots added for visual changes
- [ ] Affected documentation updated (docusaurus, tutorial, CMS)

[PBE-5004]: https://stream-io.atlassian.net/browse/PBE-5004?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ